+++ CHANGELOG.md
@@ -35,6 +35,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added GitHub Action to generate MAPL3 Ford Docs
 - Added capability for HistoryCollectionGridComp to extract field names from expressions
 - Added ability for HistoryCollectionGridComp to extract multiple field names from expressions
+- Added procedures to get information about an ESMF_FieldBundle in History3G
+- Added module for keys to ESMF_Info metadata used in MAPL3G
 
 ### Changed